requirements.txt
streamlit
tensorflow
opencv-python
numpy
Pillow
streamlit-drawable-canvas
